Fixed microsoftism

+++ math/src/scalar/mathf.cpp
@@ -157,7 +157,7 @@ F64 Mathf::Trunc( const F64 f, const F64 exp )
 
 F32 Mathf::Trunc( const F32 f, const F32 exp )
 {
-    return std::powf( f, exp );
+    return std::pow( f, exp );
 }
 
 F64 Mathf::Rint( const F64 x )
